Add command to resize final image and toggle to output gif
Since every pixel has to be replaced by an ASCII character the dimensions of the final image can become far larger than the resized original image. The issue becomes worse the larger the font size.

After converting an image to ASCII the script can now be told to resize the final image as well before saving it to output directory.